requirements.txt
asgiref==3.11.1
attrs==25.4.0
dj-database-url==3.1.2
Django==5.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
ruff==0.15.2
sqlparse==0.5.5
uritemplate==4.2.0
